requirements.txt
pymysql~=1.0.2
numpy~=1.20.1
pillow~=8.2.0
dbutils~=1.2
pyside6~=6.2.3
opencv-python~=4.5.5.62
torch~=1.11.0+cu113
scipy~=1.6.2
matplotlib~=3.3.4
tqdm~=4.59.0
pandas~=1.2.4
dlib~=19.19.0
imutils~=0.5.4
scikit-learn~=0.24.1
colorama~=0.4.4
wandb~=0.12.11
yaml~=0.2.5
pyyaml~=5.4.1
seaborn~=0.11.1
torchvision~=0.12.0+cu113
requests~=2.23.0
thop~=0.0.31-2005241907